+++ pkg/network/admitter/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "admit.go",
         "binding.go",
+        "netsource.go",
         "slirp.go",
         "validator.go",
     ],
@@ -24,6 +25,7 @@ go_test(
         "admit_suite_test.go",
         "admit_test.go",
         "binding_test.go",
+        "netsource_test.go",
         "slirp_test.go",
     ],
     deps = [
