+++ pkg/network/admitter/admit_test.go
@@ -93,18 +93,4 @@ var _ = Describe("Validating VMI network spec", func() {
 				Field:   "fake.domain.devices.interfaces[0].state",
 			}))
 	})
-
-	It("support only a single pod network", func() {
-		const net1Name = "default"
-		const net2Name = "default2"
-		vmi := v1.VirtualMachineInstance{}
-		vmi.Spec.Networks = []v1.Network{
-			{Name: net1Name, NetworkSource: v1.NetworkSource{Pod: &v1.PodNetwork{}}},
-			{Name: net2Name, NetworkSource: v1.NetworkSource{Pod: &v1.PodNetwork{}}},
-		}
-		validator := admitter.NewValidator(k8sfield.NewPath("fake"), &vmi.Spec, stubSlirpClusterConfigChecker{})
-		causes := validator.Validate()
-		Expect(causes).To(HaveLen(1))
-		Expect(causes[0].Message).To(Equal("more than one interface is connected to a pod network in fake.interfaces"))
-	})
 })
