+++ pkg/network/admitter/validator.go
@@ -53,6 +53,7 @@ func (v Validator) Validate() []metav1.StatusCause {
 	var causes []metav1.StatusCause
 
 	causes = append(causes, validateSinglePodNetwork(v.field, v.vmiSpec)...)
+	causes = append(causes, validateSingleNetworkSource(v.field, v.vmiSpec)...)
 	causes = append(causes, validateInterfaceStateValue(v.field, v.vmiSpec)...)
 	causes = append(causes, validateInterfaceBinding(v.field, v.vmiSpec)...)
 	causes = append(causes, validateSlirpBinding(v.field, v.vmiSpec, v.configChecker)...)
